+++ tests/spot/c2c/test_c2c_trade_history.py
@@ -4,7 +4,7 @@ from tests.util import random_str, timestamp
 from tests.util import mock_http_response
 from urllib.parse import urlencode
 from binance.spot import Spot as Client
-from binance.error import ParameterRequiredError, ClientError
+from binance.error import ParameterRequiredError
 
 mock_item = {"key_1": "value_1", "key_2": "value_2"}
 
@@ -22,14 +22,6 @@ params = {
 
 def test_c2c_trade_history_without_trade_type():
     """Tests the API endpoint of c2c trade history without trade type"""
-
-    params = {
-        "startTimestamp": timestamp(),
-        "endTimestamp": timestamp(),
-        "page": 1,
-        "rows": 100,
-    }
-
     client = Client(key, secret)
     client.c2c_trade_history.when.called_with("").should.throw(ParameterRequiredError)
 
@@ -45,3 +37,4 @@ def test_c2c_trade_history():
 
     client = Client(key, secret)
     response = client.c2c_trade_history(**params)
+    response.should.equal(mock_item)