+++ tests/spot/bswap/test_bswap_swap_history.py
@@ -2,7 +2,6 @@ import responses
 from binance.spot import Spot as Client
 from tests.util import mock_http_response
 from tests.util import random_str
-from binance.error import ParameterRequiredError
 
 mock_item = {"key_1": "value_1", "key_2": "value_2"}
 key = random_str()