+++ tests/spot/futures/test_futures_loan_borrow.py
@@ -4,7 +4,7 @@ from urllib.parse import urlencode
 from tests.util import random_str
 from tests.util import mock_http_response
 from binance.spot import Spot as Client
-from binance.error import ParameterRequiredError, ClientError
+from binance.error import ParameterRequiredError
 
 mock_item = {"key_1": "value_1", "key_2": "value_2"}
 mock_exception = {"code": -1105, "msg": "error message."}