+++ tests/spot/mining/test_mining_statistics_list.py
@@ -1,10 +1,9 @@
 import responses
 
-from urllib.parse import urlencode
 from tests.util import random_str
 from tests.util import mock_http_response
 from binance.spot import Spot as Client
-from binance.error import ParameterRequiredError, ClientError
+from binance.error import ParameterRequiredError
 
 mock_item = {"key_1": "value_1", "key_2": "value_2"}