+++ docs/source/CHANGELOG.rst
@@ -9,11 +9,12 @@ Unreleased
 ------------------
 * New Fiat endpoints
 
-  ** `GET /sapi/v1/fiat/orders` to query user fiat deposit and withdraw history 
-  ** `GET /sapi/v1/fiat/payments` to query user fiat payments history 
+  * ``GET /sapi/v1/fiat/orders`` to query user fiat deposit and withdraw history 
+  * ``GET /sapi/v1/fiat/payments`` to query user fiat payments history 
 
 * Fixed
-  * Typo in `margin_max_transferable`
+
+  * Typo in ``margin_max_transferable``
 
 1.3.0 - 2021-07-22
 ------------------