+++ docs/source/CHANGELOG.rst
@@ -5,16 +5,38 @@ Changelog
 Unreleased
 ----------
 
+1.5.0 - 2021-08-17
+------------------
+
+Changed
+^^^^^^^
+
+* ``GET api/v3/exchangeInfo`` now supports single or multi-symbol query
+* ``GET api/v3/myTrades`` has a new optional field ``orderId``
+
+Added
+^^^^^
+
+* ``GET /sapi/v1/c2c/orderMatch/listUserOrderHistory`` to query user C2C trade history
+
+
 1.4.0 - 2021-07-30
 ------------------
+
+Added
+^^^^^
+
+
 * New Fiat endpoints
 
   * ``GET /sapi/v1/fiat/orders`` to query user fiat deposit and withdraw history 
   * ``GET /sapi/v1/fiat/payments`` to query user fiat payments history 
 
-* Fixed
+Fixed
+^^^^^
 
-  * Typo in ``margin_max_transferable``
+
+* Typo in ``margin_max_transferable``
 
 1.3.0 - 2021-07-22
 ------------------
@@ -22,11 +44,11 @@ Unreleased
 Added
 ^^^^^
 
-* New endpoint for Wallet:
+
+* New endpoints for Wallet:
 
   * ``POST /sapi/v1/asset/get-funding-asset`` to query funding wallet, includes Binance Pay, Binance Card, Binance Gift Card, Stock Token
   * ``GET /sapi/v1/account/apiRestrictions`` to query user API Key permission
-
 
 1.2.0 - 2021-07-12
 ------------------
@@ -34,8 +56,8 @@ Added
 Changed
 ^^^^^^^
 
-* Remove default value in the parameters
 
+* Remove default value in the parameters
 
 1.1.1 - 2021-06-24
 ------------------
@@ -43,14 +65,15 @@ Changed
 Changed
 ^^^^^^^
 
-* Upgrade the dependency packages
 
+* Upgrade the dependency packages
 
 1.1.0 - 2021-06-23
 ------------------
 
 Added
 ^^^^^
+
 
 * A link to the document on ``README.md``
 * Enabled the sub menu on document nav bar.
@@ -61,11 +84,11 @@ Added
   * ``GET /sapi/v1/managed-subaccount/asset`` to query managed sub-account asset details (only for investor master account)
   * ``POST /sapi/v1/managed-subaccount/withdraw`` to withdrawal assets from the managed sub-account (only for investor master account)
 
-
 1.0.0 - 2021-06-15
 ------------------
 
 Added
 ^^^^^
+
 
 * First release, please find details from ``README.md``